% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wake.R
\name{wake_developments}
\alias{wake_developments}
\title{Get Wake County development records}
\usage{
wake_developments(status = NULL)
}
\arguments{
\item{status}{Optional plan status to filter by}
}
\value{
tibble with development records
}
\description{
Get Wake County development records
}
