% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wake.R
\name{wake_search}
\alias{wake_search}
\title{Search Wake County development records by keyword}
\usage{
wake_search(query)
}
\arguments{
\item{query}{Character string to search for}
}
\value{
tibble (same schema as wake_developments)
}
\description{
Search Wake County development records by keyword
}
